Add createGridBoard, addHoverEffect, rangeBtn
createGridBoard function changes grid settings in CSS based on user
choice. By default, it's a 16x16 grid. Also, it creates and appends the
required number of divs to create a grid on the page.

addHoverEffect function adds listener to all divs and when you hover over
them it changes their color.

rangeBtn listener runs the createGridBoard function whenever the user
changes its value. So it creates a new grid based on its choice.