Improve HoverEffect and make unselectable squares
The hover effect will now work only when the mouse button is clicked:
Added mousedown and mouseup listeners to the whole document.
I added the global variable isMouseDown with the boolean value.
Whenever mouse button is clicked isMousedown is set to true, when the button
isn't clicked it is set to false.
Hence, onmouseover listener functionality will only work when isMouseDown
is true.
I also added an onmousedown listener to the squares. So it works even with
a single click.

CSS updates: made all squares unselectable to prevent bugs.